Common url module for muuntaja/merge
